+++ src/pages/index.astro
@@ -1,6 +1,7 @@
 ---
 
 import {HeroSection} from '../section/HeroSection'
+import { NavBar } from '../section/NavBar'
 import { Headers } from '../section/Header';
 import Layout from '../layouts/Layout.astro';
 import { Vission } from '../section/Vission';
@@ -14,7 +15,9 @@ import { Footer } from '../section/Footer';
 
 ---
 
+
 <Layout >
+    <NavBar  client:load/>
     <Headers client:load/>
 	<HeroSection client:load/>
     <Vission client:load />
