+++ src/style/global.css
@@ -6,6 +6,7 @@
   body {
     background-color: #141414;
 }
+
 }
 input {
     all: unset; /* Resets all inherited and default styles */
